+++ crates/universal-wallet/macro-helpers/src/foreign_attributes.rs
@@ -5,14 +5,31 @@ use darling::FromMeta;
 use quote::ToTokens;
 use syn::{Ident, Meta};
 
+/// For parsing: list of attributes to not ignore
+const PARSED_FOREIGN_ATTRS: [&str; 2] = ["serde", "borsh"];
+/// For parsing: helper struct parsing a single attr
+#[derive(Debug, FromMeta)]
+pub enum SingleForeignAttr {
+    Serde(Serde),
+    Borsh(Borsh),
+}
+
+/// Output structure collecting the parsed information from every foreign attribute we're handling
 #[derive(Debug, Default, FromMeta)]
 pub struct ForeignAttrs {
-    serde: SerdeRename,
+    pub serde: Serde,
+    pub borsh: Borsh,
+}
+
+#[allow(dead_code)]
+#[derive(Debug, Default, FromMeta)]
+pub struct Borsh {
+    pub use_discriminant: bool,
 }
 
 #[derive(Debug, Default, FromMeta)]
 #[darling(allow_unknown_fields)]
-pub struct SerdeRename {
+pub struct Serde {
     /// used by serde on structs and variants
     rename_all: Option<WithOriginal<String, Meta>>,
     // TODO: these aren't currently used in the SDK and can be implemented later if needed
@@ -26,7 +43,7 @@ pub struct SerdeRename {
 /// then some extra logic needs to be added to construct a merged `SerdeRename` from the parent
 /// type's one (where `rename_all = Some(...)`) and the one from the field (where `rename =
 /// Some(...)`). The latter should override the former.
-impl SerdeRename {
+impl Serde {
     fn rename_using_rename_all(&self, original: &Ident) -> Result<String, syn::Error> {
         match &self.rename_all {
             Some(str) if str.parsed == "snake_case" => {
@@ -53,18 +70,26 @@ impl SerdeRename {
     }
 }
 
-pub fn parse_serde_rename_attrs(attrs: Vec<syn::Attribute>) -> darling::Result<SerdeRename> {
+pub fn parse_foreign_attrs(attrs: Vec<syn::Attribute>) -> darling::Result<ForeignAttrs> {
     let mut res = ForeignAttrs::default();
     for a in attrs.iter() {
-        if a.path().is_ident("serde") {
+        if a.path()
+            .get_ident()
+            .is_some_and(|p| PARSED_FOREIGN_ATTRS.contains(&p.to_string().as_str()))
+        {
             let meta_list = parse_attribute_to_meta_list(a)?;
             let list = NestedMeta::parse_meta_list(meta_list.into_token_stream())?;
-            let parsed_attr = ForeignAttrs::from_list(&list)?.serde;
-            // TODO: better logic here if parsing multiple attributes
-            if parsed_attr.rename_all.is_some() {
-                res.serde.rename_all = parsed_attr.rename_all;
+            let known_attr = SingleForeignAttr::from_list(&list)?;
+            match known_attr {
+                SingleForeignAttr::Serde(serde) => {
+                    // TODO: better merging logic here if parsing multiple attributes
+                    res.serde.rename_all = serde.rename_all;
+                }
+                SingleForeignAttr::Borsh(borsh) => {
+                    res.borsh.use_discriminant = borsh.use_discriminant;
+                }
             }
         }
     }
-    Ok(res.serde)
+    Ok(res)
 }